assignments/health_wellness_agent/pyproject.toml
[project]
name = "health-wellness-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.1.0",
    "pyarrow==16.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit==1.37.1",
]
